requirements.txt
fastapi
uvicorn
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
openai
pandas
python-dotenv
chromadb
PyPDF2
google-generativeai
spacy
scikit-learn
bcrypt
django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
mssql-django
pyodbc